refactor(layout): introduce (tabbed)/(default-header) Route Groups so navigation preserves layout
PR #275 マージ後、Web 版 throttling 環境でタブ切替時に Header / TabNavigation を
含む画面全体が再描画される問題が判明していた。Next.js app router の layout
システムを正しく活用し、共通 Header / TabNavigation を上位 layout で保持させて
navigation 時には main の中身だけ差し替わる構造へ refactor する。

主な変更:
- (authenticated)/(tabbed)/layout.tsx を新規追加し、TabNavigation を共有 boundary
  として常駐させる。SP 表示時は下部に固定、PC 表示時は非表示
- (authenticated)/(tabbed)/(default-header)/layout.tsx を新規追加し、DefaultHeader
  と main 構造を共有。3 page (personal/pages, personal/pages/[page_id], mypage) で
  Header と main の枠を維持しつつ children だけ差し替えられる
- AuthenticatedHeader.tsx (Client Wrapper) を新設。layout (Server) から
  DefaultHeader を呼ぶための薄いラッパで useAuth から user を取得して渡す
- DefaultHeader: dead code だった settingsHref prop を削除、showTooltip prop も
  削除し pathname (`/${locale}/personal/pages` 限定) で内部判定する形へ refactor
- SocialFeedHeader: profileImageUrl prop を削除し内部 useAuth で取得するよう
  refactor。(tabbed)/social/posts/page.tsx から children なしで呼び出せる形に

ファイル移動 (git mv で履歴維持):
- (authenticated)/personal/pages/* → (authenticated)/(tabbed)/(default-header)/personal/pages/*
- (authenticated)/personal/pages/[page_id]/page.tsx, PageDetail.tsx, page.module.css, page.test.tsx → 同上
- (authenticated)/mypage/* → (authenticated)/(tabbed)/(default-header)/mypage/*
- (authenticated)/social/posts/{page,SocialPostsFeed,SocialPostsFeedSkeleton}.tsx → (authenticated)/(tabbed)/social/posts/*
- 編集系 page (personal/pages/[page_id]/edit, new, social/posts/[post_id]/edit, new, search 等) は
  TabNavigation なし layout のため (authenticated) 直下のまま維持

各 page から layout の責務を取り除き、children だけ render する形に整理:
- personal/pages/page.tsx: <DefaultLayout> wrap を削除
- personal/pages/[page_id]/page.tsx: 同上
- mypage/MyPage.tsx: <DefaultLayout> wrap を削除、settingsHref prop も削除
- mypage/MyPageInitializer.tsx: settingsHref を MyPage に渡さない
- (tabbed)/social/posts/page.tsx: SocialBottomNav と SocialLayout DOM の inline
  再現を削除し、SocialFeedHeader を Suspense の外で render する形に
- SocialPostsFeed.tsx: 内部から SocialFeedHeader 呼び出しを削除

応急処置の巻き戻し:
- (authenticated)/loading.tsx: PR #276 で復元した Loader を再度 null に戻す。
  layout が常駐するため segment loading.tsx で Loader を出すと専用 Skeleton が
  上書きされてしまう
- (public)/loading.tsx: 同上

Skeleton の整理:
- MyPageSkeleton.tsx / .module.css: Header / TabNavigation の inline 再現を削除し
  main の中身 (UserInfoCardSkeleton 風) だけに
- SocialPostsFeedSkeleton.tsx / .module.css: 同様に Header の inline 再現を削除

Dead code 削除:
- DefaultLayout.tsx を削除 (もう誰も import していない、layout 化で不要に)
- SocialBottomNav.tsx を削除 ((tabbed)/layout が TabNavigation を提供するため不要)

検証:
- pnpm check / pnpm tsc --noEmit / pnpm test (276/276) / pnpm build すべて PASS
- 全 37 page route の ◐ Partial Prerender 化を維持
- /personal/pages → /mypage の navigation で Header と TabNavigation が再描画
  されず main だけ差し替わる構造を実現
