+++ frontend/src/app/[locale]/(authenticated)/(tabbed)/(default-header)/mypage/MyPage.tsx
@@ -4,7 +4,6 @@ import { useTranslations } from "next-intl";
 import { useCallback, useState } from "react";
 import type { UserProfile } from "@/components/features/personal/MyPageContent/MyPageContent";
 import { MyPageContent } from "@/components/features/personal/MyPageContent/MyPageContent";
-import { DefaultLayout } from "@/components/shared/layouts/DefaultLayout";
 import { SurveyModal } from "@/components/shared/SurveyModal/SurveyModal";
 import { useToast } from "@/contexts/ToastContext";
 import { getUserInfo, updateUserInfo } from "@/lib/api/client";
@@ -14,10 +13,9 @@ import { useUmamiTrack } from "@/lib/hooks/useUmamiTrack";
 
 interface MyPageProps {
   initialUser: UserProfile;
-  settingsHref: string;
 }
 
-export default function MyPage({ initialUser, settingsHref }: MyPageProps) {
+export default function MyPage({ initialUser }: MyPageProps) {
   const t = useTranslations();
   const [user, setUser] = useState<UserProfile>(initialUser);
   // サーバーから initialUser が渡るのでマウント時は false。mutation 後の refetch でのみ true
@@ -75,7 +73,7 @@ export default function MyPage({ initialUser, settingsHref }: MyPageProps) {
   };
 
   return (
-    <DefaultLayout settingsHref={settingsHref}>
+    <>
       <MyPageContent user={user} loading={loading} />
       <SurveyModal
         isOpen={isSurveyOpen}
@@ -84,6 +82,6 @@ export default function MyPage({ initialUser, settingsHref }: MyPageProps) {
         initialAgeRange={user.age_range}
         initialGender={user.gender}
       />
-    </DefaultLayout>
+    </>
   );
 }
