+++ frontend/src/app/[locale]/(authenticated)/(tabbed)/(default-header)/mypage/MyPageSkeleton.module.css
@@ -1,0 +1,26 @@
+/* UserInfoCardSkeleton.module.css と同じ寸法。Header / TabNavigation の枠は layout 側 */
+.profileSection {
+  display: flex;
+  align-items: center;
+  gap: 16px;
+  margin-bottom: 24px;
+}
+
+.userInfo {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+}
+
+.detailsSection {
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+}
+
+.detail {
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+}
