+++ frontend/src/app/[locale]/(authenticated)/(tabbed)/(default-header)/mypage/page.tsx
