+++ frontend/src/app/[locale]/(authenticated)/(tabbed)/(default-header)/personal/pages/PersonalPages.tsx
