+++ frontend/src/app/[locale]/(authenticated)/(tabbed)/(default-header)/personal/pages/[page_id]/PageDetail.tsx
