+++ frontend/src/app/[locale]/(authenticated)/loading.tsx
@@ -1,6 +1,6 @@
-// segment loading.tsx は page 内の <Suspense fallback> より外側で動くため、ここで
-// 何かを返すと page ごとに用意した専用 Skeleton が上書きされてしまう。null を返して
-// page level の fallback に委ねる
+// (tabbed)/(default-header)/layout 等の共通 layout が Header / TabNavigation を保持し
+// navigation 時に維持されるため、segment loading.tsx は何も表示せず page level の
+// <Suspense fallback> に委ねる
 export default function AuthenticatedLoading() {
   return null;
 }
