--- frontend/src/app/[locale]/(authenticated)/mypage/MyPageSkeleton.module.css
@@ -1,87 +1,0 @@
-/* DefaultLayout.module.css の DOM 構造と寸法を揃え CLS を抑える */
-.layout {
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  background: transparent;
-}
-
-/* DefaultHeader の枠 (header padding/min-height は実際のヘッダーと合わせる) */
-.headerShell {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 8px;
-  padding: 12px 16px;
-  min-height: 65px;
-  border-bottom: 0.5px solid var(--border-color);
-  background: var(--white);
-}
-
-.headerActions {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-}
-
-.contentWrapper {
-  flex: 1;
-  width: 100%;
-  display: flex;
-  justify-content: center;
-}
-
-.main {
-  flex: 1;
-  width: 100%;
-  max-width: 580px;
-  padding: 24px 16px 90px;
-  box-sizing: border-box;
-}
-
-/* UserInfoCardSkeleton.module.css と同じ寸法 */
-.profileSection {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-  margin-bottom: 24px;
-}
-
-.userInfo {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-}
-
-.detailsSection {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-}
-
-.detail {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-}
-
-/* DefaultLayout.module.css の .tabNavigation の高さ確保 */
-.tabNavigationShell {
-  margin-top: auto;
-  border-top: 0.5px solid var(--border-color);
-  background: var(--white);
-  padding: 10px 0 14px;
-  height: calc(10px + 14px + 36px);
-  box-sizing: border-box;
-}
-
-@media (min-width: 768px) {
-  .main {
-    padding: 32px 20px 40px;
-  }
-
-  .tabNavigationShell {
-    display: none;
-  }
-}
