+++ frontend/src/app/[locale]/(public)/loading.tsx
@@ -1,6 +1,5 @@
-// segment loading.tsx は page 内の <Suspense fallback> より外側で動くため、ここで
-// 何かを返すと page ごとに用意した専用 Skeleton が上書きされてしまう。null を返して
-// page level の fallback に委ねる
+// page level の <Suspense fallback> (例: SocialProfileSkeleton) で表示を担うため
+// segment loading.tsx は null を返す
 export default function PublicLoading() {
   return null;
 }
