--- frontend/src/components/shared/layouts/DefaultLayout/index.ts
@@ -1,1 +1,0 @@
-export { DefaultLayout } from "./DefaultLayout";
