+++ frontend/src/components/shared/layouts/index.ts
@@ -1,6 +1,5 @@
 export { ChatLayout } from "./ChatLayout";
 export { DefaultHeader } from "./common/DefaultHeader";
-export { DefaultLayout } from "./DefaultLayout";
 export { MinimalLayout } from "./MinimalLayout";
 export { NotLoggedInLayout } from "./NotLoggedInLayout";
 export { SocialHeader, SocialLayout } from "./SocialLayout";
